Remove pit_crew
