--- rmf_building_map_tools/pit_crew/__init__.py
@@ -1,1 +1,0 @@
-from .pit_crew import *
